phlsys_web: Extract webserver utilities out of arcyd-tester
Extract get_free_port() function and SimpleWebServer class out of
arcyd-tester into a new file phlsys_web.py. Also, add tests for those.

Test plan:
$ ./precommit.sh
